code refactoring + add getpettest and updatepet pest

+++ src/main/java/petStore/dto/Tag.java
@@ -4,13 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.experimental.Accessors;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class Tag {
-    public int id;
+    public Long id;
     public String name;
 }
